+++ packages/react/src/components/Form/Form.tsx
@@ -1,0 +1,11 @@
+import { Form as FormRoot } from '@cos-ui/primitives';
+import styled from 'styled-components';
+
+const Form = styled(FormRoot)`
+  display: flex;
+  flex-direction: column;
+  width: 58rem;
+  gap: 1.2rem;
+`;
+
+export default Form;
