+++ packages/react/src/components/Form/TextField.tsx
@@ -1,0 +1,20 @@
+import { Form } from '@cos-ui/primitives';
+import styled from 'styled-components';
+
+import { TextAreaStyle } from '../TextArea';
+import { TextInputStyle } from '../TextInput';
+import { FieldStyle } from './Field';
+
+const TextField = styled(Form.TextField)`
+  ${FieldStyle}
+
+  & > input {
+    ${TextInputStyle}
+  }
+
+  & > textarea {
+    ${TextAreaStyle}
+  }
+`;
+
+export default TextField;
